+++ src/containerapp/HISTORY.rst
@@ -4,8 +4,10 @@ Release History
 ===============
 Upcoming
 +++++++
+* 'az containerapp env update': Fix bugs in update environment.
 * Fix YAML create with user-assigned identity
 * Fix polling logic for long running operations.
+* Add `az containerapp ingress update` Command to Update Container App Ingress
 
 0.3.24
 ++++++